Split migration tests and drop _EXPECTED_TABLES
Separates the frozen legacy migrate() tests from the yoyo-dispatch tests
into two files so each can evolve independently, and promotes their
shared helpers to _migration_helpers.py. TestYoyoDispatch's three
references to the hard-coded _EXPECTED_TABLES set now compare against a
fresh schema.sql bootstrap via assert_schemas_equal, removing the drift
risk that let the table list silently fall out of sync with the schema.
